move openstack tests to handler_test package
